+++ app/gateways/db/memory/get_account_by_cpf.go
@@ -1,26 +1,25 @@
 package memory
 
 import (
-	"errors"
 	"github.com/patriciapedrosaa/transfer-me/app/domain/entities"
-	"github.com/patriciapedrosaa/transfer-me/app/domain/vos"
 )
 
-var errNotFound = errors.New("not found")
-
 func (m MemoryStorage) GetByCpf(cpf string) (entities.Account, error) {
-	account := m.storageAccount[cpf]
-	if account.cpf == "" {
-		return entities.Account{}, errNotFound
+	accounts, err := m.GetAccounts()
+	if err != nil {
+		return entities.Account{}, err
 	}
-
-	entityAccount := entities.Account{
-		AccountID: account.id,
-		Name:      account.name,
-		CPF:       vos.CPF(account.cpf),
-		Secret:    vos.Secret(account.secret),
-		Balance:   account.balance,
-		CreatedAt: account.createdAt,
+	for _, account := range accounts {
+		if string(account.CPF) == cpf {
+			return entities.Account{
+				AccountID: account.AccountID,
+				Name:      account.Name,
+				CPF:       account.CPF,
+				Secret:    account.Secret,
+				Balance:   account.Balance,
+				CreatedAt: account.CreatedAt,
+			}, nil
+		}
 	}
-	return entityAccount, nil
+	return entities.Account{}, errNotFound
 }
